+++ RandomSkunk.Results.UnitTests/Result_of_T_struct.cs
@@ -9,7 +9,6 @@ public class Result_of_T_struct
 
         result.IsFail.Should().BeTrue();
         result.IsSuccess.Should().BeFalse();
-        result.IsDefault.Should().BeTrue();
         result.Error.Should().BeSameAs(Error.DefaultError);
     }
 
@@ -22,7 +21,6 @@ public class Result_of_T_struct
 
             result.IsSuccess.Should().BeTrue();
             result.IsFail.Should().BeFalse();
-            result.IsDefault.Should().BeFalse();
             result.Value.Should().Be(1);
         }
 
@@ -34,7 +32,6 @@ public class Result_of_T_struct
 
             result.IsFail.Should().BeTrue();
             result.IsSuccess.Should().BeFalse();
-            result.IsDefault.Should().BeFalse();
             result.Error.Should().BeSameAs(error);
         }
     }