+++ RandomSkunk.Results/Maybe{T}.cs
@@ -82,11 +82,6 @@ public partial struct Maybe<T> : IResult<T>, IEquatable<Maybe<T>>
     public bool IsFail => _outcome == _failOutcome;
 
     /// <summary>
-    /// Gets a value indicating whether this is a default instance of the <see cref="Maybe{T}"/> struct.
-    /// </summary>
-    public bool IsDefault => _outcome == _failOutcome && _error is null;
-
-    /// <summary>
     /// Indicates whether the <paramref name="left"/> parameter is equal to the <paramref name="right"/> parameter.
     /// </summary>
     /// <param name="left">The left side of the comparison.</param>