+++ RandomSkunk.Results/Result{T}.cs
@@ -63,11 +63,6 @@ public partial struct Result<T> : IResult<T>, IEquatable<Result<T>>
     public bool IsFail => _outcome == _failOutcome;
 
     /// <summary>
-    /// Gets a value indicating whether this is a default instance of the <see cref="Result{T}"/> struct.
-    /// </summary>
-    public bool IsDefault => _outcome == _failOutcome && _error is null;
-
-    /// <summary>
     /// Indicates whether the <paramref name="left"/> parameter is equal to the <paramref name="right"/> parameter.
     /// </summary>
     /// <param name="left">The left side of the comparison.</param>