+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+### Removed
+
+- RandomSkunk.Results:
+    - Remove `IsDefault` property from each result type.
+
 ## [1.0.0-alpha19] - 2022-10-18
 
 [RandomSkunk.Results API](https://www.fuget.org/packages/RandomSkunk.Results/1.0.0-alpha19) ([diff](https://www.fuget.org/packages/RandomSkunk.Results/1.0.0-alpha19/lib/netstandard2.1/diff/1.0.0-alpha18/))